further performance enchancements on large worlds

+++ include/World/marchingWorld.h
@@ -26,8 +26,8 @@ namespace Hop::World
         void load(std::string filename){}
 
         void worldToTileData(float x, float y, Tile & h, float & x0, float & y0, float & s);
-        void neighourTileData(float x, float y, TileNeighbourData & nData, Tile & h, float & x0, float & y0, float & s);
-
+        void neighourTileData(double x, double y, TileNeighbourData & nData, Tile & h, double & x0, double & y0, double & s);
+        
         Tile tileType(int i, int j);
         void tileToIdCoord(int ix, int iy, int & i, int & j);
         bool updateRegion(float x, float y);
