+++ include/World/tileWorld.h
@@ -24,7 +24,7 @@ namespace Hop::World
         
         void worldToTileData(float x, float y, Tile & h, float & x0, float & y0, float & s);
         void boundsTileData(double x, double y, Tile & h, TileBoundsData & bounds, double & x0, double & y0, double & s);
-        void neighourTileData(float x, float y, TileNeighbourData & nData, Tile & h, float & x0, float & y0, float & s);
+        void neighourTileData(double x, double y, TileNeighbourData & nData, Tile & h, double & x0, double & y0, double & s);
         
         Tile tileType(int i, int j);
         void tileToIdCoord(int ix, int iy, int & i, int & j);
