+++ src/World/marchingWorld.cpp
@@ -244,13 +244,13 @@ namespace Hop::World
 
     void MarchingWorld::neighourTileData
     (
-        float x, 
-        float y, 
+        double x, 
+        double y, 
         TileNeighbourData & nData,
         Tile & h,
-        float & x0,
-        float & y0,
-        float & s
+        double & x0,
+        double & y0,
+        double & s
     )
     {
